html/semantics/embedded-content-0/media-elements/currentSrc_empty_if_no_src.html

<!doctype html>
<html>
 <head>
  <title>{audio,video}.currentSrc - empty if no source</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
 </head>
 <body>
  <p><a href="http://dev.w3.org/html5/spec/Overview.html#dom-media-currentsrc">spec reference</a></p>
  <audio id="a">
  </audio>
  <video id="v">
  </video>
  <div id="log"></div>
  <script>
test(function() {
 assert_equals(
  document.getElementById("a").currentSrc,
  "",
  "audioElement.currentSrc should be empty if there is no source");
}, "audio.currentSrc - empty if no source");

test(function() {
 assert_equals(
  document.getElementById("v").currentSrc,
  "",
  "videoElement.currentSrc should be empty if there is no source");
}, "video.currentSrc - empty if no source");
  </script>
 </body>
</html>
